requirements.txt
click==8.0.0
colorama==0.4.4
Flask==2.0.0
Flask-Cors==3.0.10
greenlet==1.1.0
importlib-metadata==4.0.1
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.15
typing-extensions==3.10.0.0
virtualenv==16.4.3
Werkzeug==2.0.1
zipp==3.4.1
